Return submit success from Submit.
Useful to get a synchronous submit result so that there is context to
log/record drops. For example, if an event is being queued, when it is
dropped due to queue full, it is not possible to get that event's
details in `onDropped` callback. With synchronous result, the dropped
event details are available.